backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pandas==2.2.3
python-dotenv==1.0.1
langchain-openai==0.3.14
langchain-core==0.3.56
openpyxl==3.1.5
pydantic==2.11.3
pymongo==4.10.1
